+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Remove deprecated REST endpoints in TaskController. (#731)
 - Move `WorkerModel` from `iexec-common` to `iexec-core-library`. (#735)
 - Move `TaskAbortCause` from `iexec-commons-poco` to `iexec-core-library`. (#736)
+- Remove deprecated methods in `iexec-core-library`. (#737)
 
 ### Dependency Upgrades